Remove hard coded school lookup.
The hardcoded school lookup when the dashboard loaded
was crashing the server if the school wasn't in
the database. Remove this in favor or looking up
the students for the teachers' schools or getting a list
of all students.

Added a Data service that will be used to pass around data
when moving from state to state.
- For now this only is populated when the main controller
  is created with the students for the user.
- This prevents us from hitting the server for every state
  change within the main state.