Carry over missing students on reports
If school year for the record is the same as the previous record,
all missing students are added to the record. New missing students
are calculated and the deltas are calculated from the previous entry.

If the school year is different from the previous record, student._ids
are matched by studentId but previous entry data is not used. All
missing entries are discarded and not added to the record.