Fix occassional visualization sizing bug
Resolves #419.

It appears the bug is from using ng-show. The main difference is
the graph is just hidden using show and removed and added using if.
I'm still not completely sure what the cause is or how to even force
a redraw, but switching to ng-if seems to fix the problem.